Moved option types of Breadcumbs extension.
